+++ imports/routes.jsx
@@ -3,17 +3,14 @@ import { Router, Route, Switch } from 'react-router';
 import { createBrowserHistory as createHistory } from 'history';
 
 // route components
-import PublicRoute from './PublicRoute';
-import PrivateRoute from './PrivateRoute';
-import Login from "./ui/Login";
-import SignUp from './ui/SignUp';
-import Create from './ui/Create';
-import Account from './ui/Account';
-import VideoPage from './ui/create/videos/VideoPage';
-import HomeVideos from './ui/create/videos/HomeVideos';
-import PageNotFound from './PageNotFound';
-import MenuPage from "./ui/MenuPage";
-import LandingPage from "./ui/LandingPage";
+import PublicRoute from './ui/PublicRoute';
+import Login from "./ui/User/Login";
+import SignUp from './ui/User/SignUp';
+import Create from './ui/Create/Create';
+import Account from './ui/Create/Account/Account';
+import VideoPage from './ui/Create/Videos/VideoPage';
+import PageNotFound from './ui/PageNotFound';
+import LandingPage from "./ui/LandingPage/LandingPage";
 
 const browserHistory = createHistory();
 
@@ -23,7 +20,6 @@ export const renderRoutes = () => (
             <PublicRoute exact path ="/" component={LandingPage}/>
             <PublicRoute exact path="/login" component={Login}/>
             <PublicRoute exact path="/signup" component={SignUp}/>
-            <PrivateRoute exact path="/menu" component={MenuPage}/>
             <Route exact path="/video" component={Create}/>
             <Route exact path="/account" component={Account}/>
             <Route exact path="/video/:videoId" component={VideoPage}/>
