+++ imports/ui/Create/Create.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import HomeVideos from "./create/videos/HomeVideos";
+import HomeVideos from "./Videos/HomeVideos";
 import {createMuiTheme, MuiThemeProvider} from "@material-ui/core/styles";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import BarTop from "./AppBar/BarTop";
