+++ imports/ui/Create/Videos/Common/Picture.jsx
